api/requirements.txt
opencv-python~=4.5.4.60
msrest~=0.6.21
Flask~=2.0.2
Werkzeug~=2.0.2
azure-cognitiveservices-vision-face~=0.6.0
Flask-Cors~=3.0.10